requirements.txt
aniso8601==9.0.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
packaging==24.1
pluggy==1.5.0
psycopg==3.2.3
PyJWT==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
tomli==2.0.2
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
